+++ tests/test_notification.py
@@ -25,9 +25,6 @@
 # licenses.
 
 import sys
-
-if ".." not in sys.path:
-    sys.path.insert(0, "..")
 import labbench as lb
 
 if __name__ == "__main__":