+++ tests/test_value.py
@@ -29,14 +29,10 @@ import importlib
 import sys
 import pandas as pd
 import numpy as np
-
-if ".." not in sys.path:
-    sys.path.insert(0, "..")
 import labbench as lb
 
 lb._force_full_traceback(True)
 
-remap = {True: "ON", False: "OFF"}
 flag_start = False